requirements.txt
langgraph==0.2.28
streamlit==1.38.0
python-dotenv==1.0.1
langchain-groq==0.3.0
langchain-core==0.3.45
pydantic==2.9.0
groq==0.9.0
httpx==0.27.0